Setup routing for lazy loading and Add the start of registerer panel
Routing was moved to the routing module.
Two modules were introduced: the registerer feature module
and the shared module for use by all future lazy loaded modules.
The registerer module contains the beginning of the panel layout and
a mock version of searching for patients which will later be done with http.